refactor: Corrigir nomes de interface da api imbd
