+++ src/app/components/movie-list/movie-list.component.ts
@@ -1,7 +1,7 @@
 import {Component, EventEmitter, Input, Output} from '@angular/core';
 import {NgForOf, NgIf} from "@angular/common";
 import {RouterLink, RouterLinkActive} from "@angular/router";
-import {SearchInterface, SaveMovieDataInterface} from "../../interfaces/movie-interface";
+import {ImdbResponseItemInterface, SaveMovieDataInterface} from "../../interfaces/imdb-interface";
 import {MoviesService} from "../../services/movies.service";
 
 @Component({
@@ -16,7 +16,7 @@ import {MoviesService} from "../../services/movies.service";
   templateUrl: './movie-list.component.html'
 })
 export class MovieListComponent {
-  @Input() movieList!: SearchInterface[];
+  @Input() movieList!: ImdbResponseItemInterface[];
   // eslint-disable-next-line @angular-eslint/no-output-native
   @Output() show = new EventEmitter<string | number>();
   loading: boolean = false;
@@ -28,7 +28,7 @@ export class MovieListComponent {
     this.show.emit(id);
   }
 
-  async addToList(item:SearchInterface): Promise<void> {
+  async addToList(item:ImdbResponseItemInterface): Promise<void> {
     const formData = {
       title: item.Title,
       year: item.Year,
