+++ src/app/services/movies.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {environment} from "../../environments/environment";
 import {HttpClient, HttpParams} from "@angular/common/http";
-import {MovieDetails, MovieInterface, SaveMovieDataInterface} from "../interfaces/movie-interface";
+import {ImdbResponseInterface, SaveMovieDataInterface, ImdbItemDetails} from "../interfaces/imdb-interface";
 import {Observable} from "rxjs";
 
 @Injectable({
@@ -14,13 +14,13 @@ export class MoviesService {
 
   constructor(private http: HttpClient) {}
 
-  async index(query:string = ''):Promise<Observable<MovieInterface>> {
+  async index(query:string = ''):Promise<Observable<ImdbResponseInterface>> {
     const options = {params: new HttpParams().set('s', query)}
 
     return this.http.get<never>(this.API_URL, options)
   }
 
-  async getMovie(id: string | number):Promise<Observable<MovieDetails>> {
+  async getMovie(id: string | number):Promise<Observable<ImdbItemDetails>> {
     const options = {params: new HttpParams().set('i', id)}
 
     return this.http.get<never>(this.API_URL, options)
